Feat: 자동채굴 가능
