+++ server/package.json
@@ -19,12 +19,12 @@
   "description": "",
   "main": "main.js",
   "scripts": {
-    "start": "node network/main.js",
+    "start": "node --trace-deprecation network/main.js",
     "backend": "nodemon network/main.js",
     "backend-peer": "export P2P_PORT=7002 && export HTTP_PORT=4002 && nodemon network/main.js",
     "dev": "concurrently \"npm run backend\" \"npm run start --prefix ../client\"",
     "dev-peer": "concurrently \"npm run backend-peer\" \"npm run dev-peer --prefix ../client\"",
-    "work": "node scripts/average-work.js"
+    "work": "node --trace-deprecation scripts/average-work.js"
   },
   "author": "",
   "license": "ISC",
